requirements.txt
defusedxml==0.4.1
Django==1.7.2
django-allauth==0.23.0
django-autoslug==1.8.0
django-cleanup==0.4.0
django-crispy-forms==1.5.2
django-durationfield==0.5.2
django-localeurl==2.0.2
django-multiselectfield==0.1.3
django-registration-redux==1.2
django-slim==0.7.5
feedparser==5.2.1
lorem-ipsum-generator==0.3
oauthlib==1.0.3
pif==0.7.1
Pillow==3.0.0
python3-openid==3.0.7
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
transliterate==1.7.4
wheel==0.24.0
